Move helper functions to conftest.py
