+++ tests/__init__.py
@@ -1,4 +1,3 @@
 import sys
 from os.path import abspath, dirname, join
 sys.path.append(dirname(dirname(abspath(__file__))))
-sys.path.append(join(dirname(__file__), 'helpers'))
