--- tests/helpers/utils.py
@@ -1,10 +1,0 @@
-def spoof_input(monkeypatch, input_list):
-    """ Given a list of user input values, spoofs the input() function to to iterate over each item
-        in the list with each successive call.
-
-        Parameters:
-            monkeypatch(MonkeyPatch): Monkeypatch from pytest.
-            input_list(List<str>): List of input values to iterate over.
-    """
-    gen = (value for value in input_list)
-    monkeypatch.setattr('builtins.input', lambda x: next(gen))
